requirements.txt
Django<3.0
-e git+https://github.com/mysociety/django-popolo.git#egg=django_popolo
django-autoslug
six
